refactor(sdk): move test utils directory
